MINOR: quic-be: Mux initialization
Reset the flags which denotes that the connection is waiting for a TLS handshake
completion. This must be accomplished before the mux creation.
Check the alpn has been successfully negotiated.
Then, finally create the mux. Note that the ->mux_proto of the server must be
set before calling conn_create_mux().